+++ src/main/java/mate/academy/service/MasterService.java
@@ -1,5 +1,6 @@
 package mate.academy.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import mate.academy.model.Master;
 import mate.academy.model.Order;
@@ -11,5 +12,5 @@ public interface MasterService {
 
     List<Order> getMasterOrdersById(Long id);
 
-    double getMasterSalaryById(Long id);
+    BigDecimal getMasterSalaryById(Long id);
 }
